+++ ESP32/include/Servo.h
@@ -12,10 +12,10 @@ class Servo {
         virtual bool home(void);
         virtual bool enable(void);
         virtual bool disable(void);
-        virtual bool move_to(double position, bool blocking = false);
-        virtual void move_to_slow(double position);
-        virtual double get_min_pos(void);
-        virtual double get_max_pos(void);
+        virtual bool move_to(float position, bool blocking = false);
+        virtual void move_to_slow(float position);
+        virtual float get_min_pos(void);
+        virtual float get_max_pos(void);
         virtual void periodic_task_func(void);
         State get_state(void) {
             return _state;
@@ -38,7 +38,7 @@ class Servo {
             }
         }
     protected:
-        double _curr_pos = 0.0;
+        float _curr_pos = 0.0;
         bool _curr_pos_valid = false;
         bool _locking_blocked = false;
         State _state = State::Disabled;
